requirements.txt
bcrypt~=4.0.1
celery[redis]
cryptography==43.0.1
fastapi==0.115.2
httpx
itsdangerous
Jinja2==3.1.6
motor
pydantic~=2.7.3
pydantic[email]
PyJWT~=2.10.1
pymongo
python-dotenv==1.0.1
python-multipart
pytz==2024.1
pyyaml~=6.0
redis~=4.3.4
shortuuid==1.0.12
starlette-context
uvicorn
